+++ addons/antennas/script_component.hpp
@@ -1,0 +1,3 @@
+#define COMPONENT antennas
+#define COMPONENT_BEAUTIFIED Antennas
+#include "\z\avo\addons\main\script_mod.hpp"